Move list of comments to a dedicated component + minor refactoring
